rename "keymap.h" to "layout.h" to prevent compile problems
