+++ keyboards/royal_kludge/rk61/keymaps/default/keymap.c
@@ -16,7 +16,7 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
 
 #include QMK_KEYBOARD_H
-#include "keymap.h"
+#include "layout.h"
 
 enum layer_names {
     _BASE,   // default layer
