+++ keyboards/royal_kludge/rk61/keymaps/iso-neo-l3l4/keymap.c
@@ -17,7 +17,7 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 #include QMK_KEYBOARD_H
 #include "keymap_german.h"
-#include "../default/keymap.h"
+#include "../default/layout.h"
 
 enum layer_names {
     _BASE,   // default layer
